refactored likes module
