+++ app/src/main/AndroidManifest.xml
@@ -30,7 +30,7 @@
             android:exported="true"
             android:windowSoftInputMode="adjustResize" />
         <activity
-            android:name=".feed.view.LikesActivity"
+            android:name=".likes.view.LikesActivity"
             android:exported="false" />
         <activity
             android:name=".notificationfeed.view.NotificationFeedActivity"
