+++ app/src/main/java/com/likeminds/feedsx/likes/adapter/LikesScreenAdapter.kt
@@ -1,7 +1,7 @@
-package com.likeminds.feedsx.feed.view.adapter
+package com.likeminds.feedsx.likes.adapter
 
-import com.likeminds.feedsx.feed.view.adapter.databinder.ItemLikesScreenViewDataBinder
-import com.likeminds.feedsx.feed.model.LikeViewData
+import com.likeminds.feedsx.likes.adapter.databinder.ItemLikesScreenViewDataBinder
+import com.likeminds.feedsx.likes.model.LikeViewData
 import com.likeminds.feedsx.utils.customview.BaseRecyclerAdapter
 import com.likeminds.feedsx.utils.customview.ViewDataBinder
 import com.likeminds.feedsx.utils.model.BaseViewType
